Fix issue where entries get overwritten
Fixes an issue where if multiple items were added on
wearos when phone client wasn't connected, they were
getting overwritten with the latest added entry.
This was because I was adding items to a single path.

Fix was to define a unique path for each item. Using
current millis for that which should be sufficient as
cannot add more than 1 item every millisecond.