Generate version code based on version name
Play Store doesn't allow uploading apps for phone and Wear
with the same version code even if they're for different devices.

So automatically generating them now, based on version name
and if the app is phone or wear.

Also, had to set compile SDK version back to 34 because the
wear OS dependencies require that. Will have to see if
Play Store accepts that or not. It probably should because the
requirement is only that Target SDK should be 33 or lower.