Fix add from share intent bug
Noticed that the add screen would be shown again
when pressing cancel or ok on the add dialog.
That was because I was navigating up from closing the
dialog and the receivedText passed into the nav graph
was still there, and journal entry screen was in a way
receiving it again as it recomposed causing navigation to
add to trigger again.