Use expanding text field, use dialog destination
The add entry text field now expands as you type which is a better
experience.
Also made the tag row a FlowRow so all options are visible on start

Finally, the weird animation when clicking add button was from
the destination not being declared as a dialog. Using dialog
fixed the delay and the animation