Attempt to fix serialization error
From last release, noticed that the error was
```
Unable to create @Body converter for java.util.List<DayGroup>
parameter #1 for method a.a
```

Which seems to be from proguard minification. Noticed that
didn't have JsonClass annotation on DayGroup which might be the
issue. So added that to see if it fixes the issue.