Fix crash when applying tag sort
The crash was happening because code was assuming that
tag on a journal entry will always have a non null value
which is not necessary.

Adding more than 1 value, one with tag and one without a tag
would trigger this.

Fix was to assign a NO_TAG value to entries without any tag
so that sorting can be applied in a way that no tag values are
always at the top.