Launch single instance of app
Launches a single instance of the app now even when
sharing from different apps. Also now only accepts text
sharing content because that's the only thing the app
understands.

For single instance, have to listen for changes in onNewIntent,
the idea for which was copied from here - https://funkymuse.dev/posts/compose-on-new-intent/