Fix wear OS bugs
Was posting to wear sharing potentially multiple entries with same
Id so they were not all getting to the phone.
Was also potentially exiting from saving entries if there were a
lot, earlier than they were posted to wear sharing which might
be causing not all of them to be posted.
Lastly, showing a toast if unable to add from template from tile