Fix incorrect date selected on when app opens (#166)
On opening the app, was selecting the wrong date to show by default.
This was because I was getting the difference between days as
DatePeriod and then using the number of days in the period to
determine the closest one to today. But that only takes into account
the days irrespective of number of months in the period.

For example, following periods are considered the same in my old way
(1month, 1day), (2months, 1 day), (3months, 1day)

Fixed to use `daysUntil` which gives the expected number